Move files, remove unused files
