--- .gitattributes
@@ -1,1 +1,0 @@
-*.ipynb filter=strip-notebook-output
